Rename missing_manager and model_manager to align with other file names
