+++ modelstore/models/keras.py
@@ -1,22 +1,21 @@
-#    Copyright 2020 Neal Lathia
-
+#    Copyright 2021 Neal Lathia
+#
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
 #    You may obtain a copy of the License at
-
+#
 #        http://www.apache.org/licenses/LICENSE-2.0
-
+#
 #    Unless required by applicable law or agreed to in writing, software
 #    distributed under the License is distributed on an "AS IS" BASIS,
 #    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
-
 import os
 from functools import partial
 
 from modelstore.models.common import save_json
-from modelstore.models.modelmanager import ModelManager
+from modelstore.models.model_manager import ModelManager
 from modelstore.storage.storage import CloudStorage
 
 MODEL_CONFIG = "model_config.json"
