+++ tests/models/test_missing_manager.py
@@ -11,9 +11,8 @@
 #    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
-
 import pytest
-from modelstore.models.missingmanager import MissingDepManager
+from modelstore.models.missing_manager import MissingDepManager
 
 # pylint: disable=redefined-outer-name
